This set of packages is generated by performing the following steps:
*) Run '$FUCHSIA_ROOT/lib/flutter/bin/flutter precache' to populate //lib/flutter/bin/cache
as flutter's pubspec.yaml refers to it by path
*) Run 'importer.py $FUCHSIA_ROOT/lib/flutter/packages/flutter $FUCHSIA_ROOT/apps/modular/handler'
to populate the dependencies of Flutter and modular's handler.

This commit also moves the Fuchsia LICENSE and AUTHORS files into the importer/
directory as that directory contains all of the code we have authored. Each
subdirectory of pub/ contains a LICENSE files with the license for code in that
package, each of which is either BSD, MIT, or Apache.

Change-Id: I4fef497f4fdbb3c5eabac661b684920ab06a8a3d